requirements.txt
pandas
textblob
requests
python-dotenv